Restructured for Gradle, and improved welcome screen.
