+++ .gitignore
@@ -34,3 +34,4 @@ hs_err_pid*
 
 # local Gradle files
 .gradle/
+.metadata/
