+++ src/main/java/util/GridBagConstraintsObject.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package util;
+package main.java.util;
 
 import java.awt.GridBagConstraints;
 import java.awt.Insets;
